#12 refactor : 패키지 구조 변겅
* global -> infrastructure
