+++ src/main/java/com/example/tennismate/infrastructure/config/QuerydslConfig.java
@@ -1,4 +1,4 @@
-package com.example.tennismate.global.config;
+package com.example.tennismate.infrastructure.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
