+++ src/test/java/com/example/tennismate/member/service/MemberServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.tennismate.member.service;
 
-import com.example.tennismate.global.exception.custom.DuplicatedException;
-import com.example.tennismate.global.exception.errorcode.ErrorCode;
+import com.example.tennismate.infrastructure.exception.custom.DuplicatedException;
+import com.example.tennismate.infrastructure.exception.errorcode.ErrorCode;
 import com.example.tennismate.member.application.port.MemberRepositoryPort;
 import com.example.tennismate.member.application.service.impl.MemberServiceImpl;
 import com.example.tennismate.member.dto.request.MemberRegisterRequest;
